Extract email message reader
