+++ app/apps/email/components/email-client-format.ts
@@ -1,0 +1,29 @@
+export function formatDate(value: string) {
+  if (!value) return '';
+  const date = new Date(value);
+  if (!Number.isFinite(date.getTime())) return value;
+  return new Intl.DateTimeFormat(undefined, { dateStyle: 'medium', timeStyle: 'short' }).format(date);
+}
+
+export function formatRecipients(value: string[] | string | undefined) {
+  if (!value) return '';
+  if (Array.isArray(value)) return value.filter(Boolean).join(', ');
+  return value;
+}
+
+export function extractEmailAddressForCompose(value: unknown): string {
+  if (!value) return '';
+  if (typeof value === 'string') {
+    const match = value.match(/<([^<>@\s]+@[^<>@\s]+)>/u) || value.match(/([A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,})/iu);
+    return (match?.[1] || '').trim().toLowerCase();
+  }
+  if (typeof value === 'object') {
+    const record = value as {
+      address?: unknown;
+      email?: unknown;
+      emailAddress?: { address?: unknown };
+    };
+    return extractEmailAddressForCompose(record.emailAddress?.address || record.address || record.email);
+  }
+  return '';
+}
